pyproject.toml
[project]
name = "bioml"
version = "0.1.0"
description = "Multimodal biometric ML infrastructure (iris + fingerprint) with MLOps foundations."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Amit Shukla"}]
dependencies = [
  "hydra-core>=1.3.2",
  "omegaconf>=2.3.0",
  "numpy>=1.25.2,<2.6.0",
  "pandas>=2.0",
  "pyarrow>=14.0.0",
  "Pillow>=10.0.0",
  "tqdm>=4.66",
  "torch>=2.2.0",
  "torchvision>=0.17.0",
  "tensorboard>=2.14",
]

[project.optional-dependencies]
ray = ["ray>=2.9.0"]
azure = ["azure-storage-blob>=12.20.0"]
dev = [
  "pytest>=7.4",
  "pytest-cov>=4.1",
  "ruff>=0.4.0",
  "black>=24.2.0",
  "mypy>=1.8.0",
  "types-Pillow",
  "pre-commit>=3.6.0",
]

[tool.black]
line-length = 100
target-version = ["py310"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]
select = ["E", "F", "I", "B", "UP", "N", "SIM", "RUF"]
ignore = ["E501"]
fix = true

[tool.ruff.lint.isort]
known-first-party = ["bioml"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true
ignore_missing_imports = true
pretty = true
exclude = ["^docs/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')"
]
